Add secrets in the config panel.
