+++ src/components/secrets/SecretForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SecretForm";
