+++ src/scss/_utils.scss
@@ -89,3 +89,7 @@
 .p-contextual-menu__dropdown.p-contextual-menu__dropdown--over-panel {
   z-index: z("zelda");
 }
+
+.p-modal--min-width .p-modal__dialog {
+  min-width: 50%;
+}
